Revert JobEvent and update BuildEvent
