+++ gitlab4j-models/src/test/resources/org/gitlab4j/models/job-event.json
@@ -10,19 +10,6 @@
   "job_status": "created",
   "job_allow_failure": false,
   "job_failure_reason": "script_failure",
-  "build_id": 1977,
-  "build_name": "test",
-  "build_stage": "test",
-  "build_status": "created",
-  "build_created_at": "2021-02-23T02:41:37.886Z",
-  "build_started_at": "2021-02-23T02:41:37.886Z",
-  "build_finished_at": "2021-02-23T02:41:37.886Z",
-  "build_duration": 2224,
-  "build_queued_duration": 1095.588715,
-  "build_allow_failure": false,
-  "build_failure_reason": "script_failure",
-  "retries_count": 2,
-  "pipeline_id": 2366,
   "project_id": 380,
   "project_name": "gitlab-org/gitlab-test",
   "user": {
@@ -32,15 +19,11 @@
   },
   "commit": {
     "id": 2366,
-    "name": "Build pipeline",
     "sha": "2293ada6b400935a1378653304eaf6221e0fdb8f",
     "message": "test\n",
     "author_name": "User",
     "author_email": "user@gitlab.com",
-    "status": "created",
-    "duration": 2264.0,
-    "started_at": "2025-04-09T17:00:09.082Z",
-    "finished_at": "2025-04-09T18:00:09.082Z"
+    "status": "created"
   },
   "repository": {
     "name": "gitlab_test",
@@ -49,28 +32,5 @@
     "git_ssh_url": "git@192.168.64.1:gitlab-org/gitlab-test.git",
     "git_http_url": "http://192.168.64.1:3005/gitlab-org/gitlab-test.git",
     "visibility_level": 20
-  },
-  "project":{
-    "id": 380,
-    "name": "Gitlab Test",
-    "description": "Atque in sunt eos similique dolores voluptatem.",
-    "web_url": "http://192.168.64.1:3005/gitlab-org/gitlab-test",
-    "git_ssh_url": "git@192.168.64.1:gitlab-org/gitlab-test.git",
-    "git_http_url": "http://192.168.64.1:3005/gitlab-org/gitlab-test.git",
-    "namespace": "Gitlab Org",
-    "visibility_level": 20,
-    "path_with_namespace": "gitlab-org/gitlab-test",
-    "default_branch": "master"
-  },
-  "runner": {
-    "active": true,
-    "runner_type": "project_type",
-    "is_shared": false,
-    "id": 380987,
-    "description": "shared-runners-manager-6.gitlab.com",
-    "tags": [
-      "linux",
-      "docker"
-    ]
   }
 }
